fix(build): move TestTagBlockError out of "use server" file
Next.js's "use server" directive restricts module exports to async
functions. The TestTagBlockError class export I introduced in Wave 1.7
(commit f4e60ac) silently broke `npm run build` — Turbopack failed
with 37 errors all cascading from src/server/warehouse.ts:57. Caught
by Track D's build run; missed by Wave 1's verification because tests
and tsc don't enforce Next.js's server-action constraint.

Fix: move the class to src/lib/errors/test-tag-block-error.ts (a
plain TS module). warehouse.ts now imports it. No runtime behavior
change — the class is identical.

Build: green. Tests: 1724/1724. Type-check: 13 pre-existing pdfme
errors, no new ones.
